rename: DetailNotification => WriteNotification
공고글 작성페이지 이기 때문에 수정
