--- src/components/templates/DetailNotificationTemplate.tsx
@@ -1,12 +1,0 @@
-import * as S from '../../styles/templates/DetailNotificationTemplate';
-import DetailNotification from '../organisms/DetailNotification';
-
-const DetailNotificationTemplate = () => {
-  return (
-    <S.Container>
-      <DetailNotification />
-    </S.Container>
-  );
-};
-
-export default DetailNotificationTemplate;
