+++ src/components/templates/WriteNotificationTemplate.tsx
@@ -1,0 +1,12 @@
+import * as S from '../../styles/templates/WriteNotificationTemplate';
+import WriteNotification from '../organisms/WriteNotification';
+
+const WriteNotificationTemplate = () => {
+  return (
+    <S.Container>
+      <WriteNotification />
+    </S.Container>
+  );
+};
+
+export default WriteNotificationTemplate;
