--- src/pages/DetailNotificationPage.tsx
@@ -1,7 +1,0 @@
-import DetailNotificationTemplate from '../components/templates/DetailNotificationTemplate';
-
-const DetailNotificationPage = () => {
-  return <DetailNotificationTemplate />;
-};
-
-export default DetailNotificationPage;
