+++ src/pages/WriteNotificationPage.tsx
@@ -1,0 +1,7 @@
+import WriteNotificationTemplate from '../components/templates/WriteNotificationTemplate';
+
+const DetailNotificationPage = () => {
+  return <WriteNotificationTemplate />;
+};
+
+export default DetailNotificationPage;
