+++ src/styles/organisms/WriteNotification.ts
