+++ src/styles/templates/WriteNotificationTemplate.ts
